requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
langchain-pinecone==0.1.3
pinecone==4.0.0
langchain-community==0.2.4
langchain==0.2.11
langchain-text-splitters==0.2.1
transformers==4.43.3
bitsandbytes==0.40.2
accelerate==0.21.0
fitz
pptx
pymupdf
unstructured
openai